Fault Detection + End of Execution Synchronization
Used select + return of send function so the sender can detect the failure. In a similar way, used the return of the recv function so that the receiver can detect the failure.

Also added a function on the receiver that sends a "finished" message, telling the previous GPU that it has already finished it's execution.

Also added a function on the sender that does not kill the socket right away, bute waits until it receives the finished message from the next GPU, which means that it has already finished it's execution.

These two functions are used alongside each other in order to detect the failure in a GPU, even when the previous GPU already finished it's execution.