Controller - Identify and react to Failure - First READ
Implemented a polling mechanism that checks for the existance of both:  a message READ on the socket as well as the failure.txt file in order to detect failure or progress of execution. Currently it is implemented just on the inner loop READ signal.

Also included the killcudalign() function that creates an executes a command for each GPU to kill it's cudalign instance.