Failure detection every READ and END
Inserted the failure detection code on every READ and END, so the controller can detect the failure at any given time.

SocketCellsReader: added the MSG_NOSIGNAL to avoid getting a broken pipe error;  Also implemented a function to delete files of past executions, which were compromising the failure detection and signalization of cudalign.

SocketCellsWriter: added the call to the failuresignal function when SCW detects failure, right after it finishes to send it's GPU cells.