Failure Detection pre-execution
Tries to identify the available GPU for the MASA-CUDAlign execution based on both the configuration file and on the retries to the socket connection.

Tries to connect to socket and sleeps for 3 times. If it still does not connect, tha execution continues without the failed GPU and the parameters are updated.