Close balancers connections
Closing terminals and the connections between controller and balancers.

Added a sleep at the end of the balancer so CUDAlign has the appropriate time to save it's data structures before the end of balancer's execution -> CUDAlign's execution

Added calls to the function signalfailure whenever SCR/SCW cannot even start a connection with the other GPU's socket.