Finish Confirmation Protocol
Implemented a function (controller) that, after identifying the dynend file, deletes it and waits for another one to be created or timeout. CUDAlign creates this dynend file, deletes the sockets and then creates the dynend file again.

The goal of this implementation is to avoid the cenario in which the fail occurs after cudalign has written dynend, but is still deleting the socket.