Execution with fewer GPU
If a GPU fails and does not come back, some parameters must be updated. This was my first modification: a function that updates the value of these parameters. One of this parameters is the number of parts, which must be reduced in order for them to be divided equally among the reamining GPUs.

It is important to note that, in order for a intermedium iteration to execute properly, it must start from where the previous finished. Since at each iteration the matrix is divided according to the split values, If a fail occurs, I must keep the matrix proportions. Because of this, I created a function that is called upon the unsuccess of connect agents. It takes the sum of splits of the first breakpoint and makes it constant for the next iteration, dividing it equally between the remaining GPUs.

Marco used (part_number%number_of_GPUs) in order to check if that GPU was the last, but this technique does not apply if the number of active GPU changes and any breakpoints were already processed. Because of this, to check if a GPU is the last, I check if it flushes columns to a socket. Otherwise it either writes the last column to disk, as a checkpoint (breakpoint), in which case it is the last GPU of a iteration; or does not write at all, in each case it is the right-most GPU.

Marco also used the part number and the amount of GPUs in other contexts and, because of this, I needed to reimplement, for exemaple, how to identify the first GPU.

In CUDAlign, I implemented a lastgpu flag, which is declarated in the extern.hpp file, created by Marco. It's value is modified in sw_stage1.cpp and on URLCellsWriter.cpp following the same logic of checking if the flush value is not socket.