Controller (part) and CUDAlign (file detection) Updates
-> Controller:
Updated how the part variable is updated so it that it always has the value of the current iteration. Previously, as soon as the command was sent,the value of part was incremented to match the next iteration.

-> CUDAlign:
Removed the file detection requisite in order for the SCR to detect failure.

The problematic cenario was that the first GPU finished it's execution and wrote the signal file, but then failed. The next GPU detected the failure, but did not do anything because the signal file was already created. Since the socket closure is now the last thing done in execution, there's no need to check the existance of this signal file.