Last Iteration Execution inside Loop + Dynend file write correction
-> Controller:
Integrated last iteration inside the loop, since if a fail occurs in the last iteration and a GPU does not come back, the command to each GPU must change.
Changed name of part character variable so it has the same name all along the code.

-> CUDAlign
correction: just the last GPU writes the dynend file.