Check Breakpoint's name + Read config gile
This version was the one used in the first documented tests (Tempo de Execução: FT-dynBP x dynBP file)

The breakpoint's path is initialized with "unavailable".

Added some lines of code to check the breakpoint's path is indeed unavailable. If iti is indeed,  instead of trying to open it, I just return 0, saying that the breakpoint does not exist.

Also added a call to the read configuration file function everytime a fail occur.