Deleting previous dirs + New dynenf file + Customizable variables
Deleted some printfs in CUDAlign and Controller.

Added a command to delete failed iteration folders.

Added a new file to the finish confirmation function to avoid possibility of deadlocks (did'nt happen but could)

Added myIP and username variables so that the user can cheange them easier, according to it's environemnt.